app/requirements.txt
torch==1.11.0
grad-cam==1.3.7
fastapi==0.77.1
uvicorn==0.17.6
python-multipart==0.0.5
opencv-python==4.5.5.64
tensorflow==2.8.0
cvlib==0.2.7
matplotlib==3.5.2
torchvision==0.12.0
tqdm==4.64.0
scikit_learn==1.1.0
torchinfo==1.6.6
pandas==1.4.2
Pillow==9.1.1
pymongo==4.1.1
